Updating build system, removing C/C++ runtime. External: Removing freetype and replacing it with stb_truetype, the main reason is that stb_truetype doesn't requires the C runtime. General: Better code quality.

+++ src/2d/animated_sprite_2d.cpp
@@ -14,9 +14,9 @@ void AnimatedSprite2D::deinit()
 	data.current_animation.destroy();
 }
 
-void AnimatedSprite2D::internal_update(f64 dt)
+void AnimatedSprite2D::internal_update(f32 dt)
 {
-	data.remain -= dt;
+	data.remain -= f32(dt);
 
 	if (data.remain > 0)
 		return;
