+++ src/canvas/label.cpp
@@ -21,7 +21,7 @@ void Label::render()
 	Transform2D transform = get_global_transform();
 	const Vector2 scale = transform.get_scale();
 	const Vector2 begin_pos = transform.get_position();
-	const Vector2 font_size = Vector2(get_font_size());
+	const Vector2 font_size = Vector2(f32(get_font_size()));
 
 	const Font::FontTheme& font_theme = data.font->get_font_theme(get_font_size());
 	
@@ -52,7 +52,9 @@ void Label::render()
 			RenderCommand::CanvasFlags(RenderCommand::FLAG_CANVAS_FLIP_V | RenderCommand::FLAG_CANVAS_FONT_CHAR)
 		);
 
-		transform.translate(Vector2(glyph.advance.x, 0));
+		transform.translate(
+			Vector2(f32(glyph.advance.x), 0)
+		);
 	}
 }
