+++ src/canvas/progress_bar.cpp
@@ -10,7 +10,7 @@ void ProgressBar::init(const CreateInfo&)
     mark(MARK_INTERNAL_UPDATE);
 }
 
-void ProgressBar::internal_update(f64 dt)
+void ProgressBar::internal_update(f32 dt)
 {
     data.value = math::move_to(data.value, data.current_value, update_speed * dt);
 }
