+++ src/object/timer.h
@@ -24,7 +24,7 @@ struct Timer : Object
     void init(const CreateInfo&);
     
     void enter();
-    void internal_update(f64 dt);
+    void internal_update(f32 dt);
 
     void start();
     void stop();
