+++ src/scene/scene_manager.cpp
@@ -25,7 +25,7 @@ void SceneManager::initialize(mem::Allocator allocator)
     data.current_scene = nullptr;
     data.current_camera = nullptr;
     
-    data.last_time = Time::get_time();
+    data.last_time = f32(Time::get_time());
     data.time_acum = 0;
     data.delta_time = 0;
     
@@ -84,7 +84,7 @@ void SceneManager::step()
 
     _handle_change_scene();
 
-    f64 current = Time::get_time();
+    f32 current = f32(Time::get_time());
     data.delta_time = current - data.last_time;
     data.last_time = current;
